Add new scenario for camparison between Basic vs. Cbap Only schedulers (#21)
* Add evaluation scenario for scheduling algorithms

Adding new codebooks generated with MATLAB.
Tapering (kaiser window) has been used to obtain these specific configurations and, in general, to obtain codebook with a taget size.
Codewords have been normalized to unit norm.
